Agent Option Tests: move mocking to utility
